+++ nugu-agent/src/main/java/com/skt/nugu/sdk/agent/display/ControlScrollDirectiveHandler.kt
@@ -134,7 +134,7 @@ class ControlScrollDirectiveHandler(
     override fun getConfiguration(): Map<NamespaceAndName, BlockingPolicy> {
         val blockingPolicy = BlockingPolicy(
             BlockingPolicy.MEDIUM_AUDIO,
-            true
+            BlockingPolicy.MEDIUM_AUDIO_ONLY
         )
 
         val configuration = HashMap<NamespaceAndName, BlockingPolicy>()