+++ nugu-agent/src/main/java/com/skt/nugu/sdk/agent/utility/UtilityAgent.kt
@@ -1,0 +1,73 @@
+/**
+ * Copyright (c) 2020 SK Telecom Co., Ltd. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http:www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.skt.nugu.sdk.agent.utility
+
+import com.google.gson.JsonObject
+import com.skt.nugu.sdk.agent.utility.handler.BlockDirectiveHandler
+import com.skt.nugu.sdk.agent.version.Version
+import com.skt.nugu.sdk.core.interfaces.capability.CapabilityAgent
+import com.skt.nugu.sdk.core.interfaces.common.NamespaceAndName
+import com.skt.nugu.sdk.core.interfaces.context.*
+import com.skt.nugu.sdk.core.interfaces.directive.DirectiveSequencerInterface
+import com.skt.nugu.sdk.core.utils.Logger
+
+class UtilityAgent(
+    contextManager: ContextManagerInterface,
+    directiveSequencer: DirectiveSequencerInterface
+) : CapabilityAgent, SupportedInterfaceContextProvider {
+    companion object {
+        private const val TAG = "UtilityAgent"
+        const val NAMESPACE = "Utility"
+        val VERSION = Version(1, 0)
+
+        private fun buildCompactContext(): JsonObject = JsonObject().apply {
+            addProperty("version", VERSION.toString())
+        }
+
+        private val COMPACT_STATE: String = buildCompactContext().toString()
+    }
+
+    override fun getInterfaceName(): String = NAMESPACE
+
+    private val contextState = object : BaseContextState {
+        override fun value(): String = COMPACT_STATE
+    }
+
+    init {
+        contextManager.setStateProvider(namespaceAndName, this)
+        provideState(contextManager, namespaceAndName, ContextType.FULL, 0)
+        provideState(contextManager, namespaceAndName, ContextType.COMPACT, 0)
+
+        directiveSequencer.addDirectiveHandler(BlockDirectiveHandler())
+    }
+
+    override fun provideState(
+        contextSetter: ContextSetterInterface,
+        namespaceAndName: NamespaceAndName,
+        contextType: ContextType,
+        stateRequestToken: Int
+    ) {
+        Logger.d(TAG, "[provideState] namespaceAndName: $namespaceAndName, contextType: $contextType, stateRequestToken: $stateRequestToken")
+        contextSetter.setState(
+            namespaceAndName,
+            contextState,
+            StateRefreshPolicy.NEVER,
+            contextType,
+            stateRequestToken
+        )
+    }
+}