+++ nugu-android-helper/src/main/java/com/skt/nugu/sdk/platform/android/NuguAndroidClient.kt
@@ -67,6 +67,7 @@ import com.skt.nugu.sdk.agent.system.SystemAgentInterface
 import com.skt.nugu.sdk.agent.text.TextAgentInterface
 import com.skt.nugu.sdk.agent.tts.TTSAgentInterface
 import com.skt.nugu.sdk.agent.tts.handler.StopDirectiveHandler
+import com.skt.nugu.sdk.agent.utility.UtilityAgent
 import com.skt.nugu.sdk.client.ClientHelperInterface
 import com.skt.nugu.sdk.client.NuguClient
 import com.skt.nugu.sdk.client.NuguClientInterface
@@ -205,6 +206,7 @@ class NuguAndroidClient private constructor(
         internal var defaultDisplayDuration = 7000L
 
         internal var textSourceHandler: TextAgentInterface.TextSourceHandler? = null
+        internal var textRedirectHandler: TextAgentInterface.TextRedirectHandler? = null
 
         internal var enableDisplay: Boolean = true
         internal var enableLocation: Boolean = true
@@ -337,6 +339,12 @@ class NuguAndroidClient private constructor(
          */
         fun textSourceHandler(handler: TextAgentInterface.TextSourceHandler?) =
             apply { this.textSourceHandler = handler }
+
+        /**
+         * @param handler the handler for text redirect directive. If not provided, default behavior at TextAgent.
+         */
+        fun textRedirectHandler(handler: TextAgentInterface.TextRedirectHandler?) =
+            apply { this.textRedirectHandler = handler }
 
         /**
          * @param enable the flag to enable or disable display.
@@ -676,7 +684,8 @@ class NuguAndroidClient private constructor(
                         getMessageSender(),
                         getContextManager(),
                         getDialogAttributeStorage(),
-                        builder.textSourceHandler
+                        builder.textSourceHandler,
+                        builder.textRedirectHandler
                     ).apply {
                         getDirectiveSequencer().addDirectiveHandler(this)
                     }
@@ -813,6 +822,13 @@ class NuguAndroidClient private constructor(
             builder.clientVersion?.let {
                 clientVersion(it)
             }
+
+            addAgentFactory(UtilityAgent.NAMESPACE, object: AgentFactory<UtilityAgent> {
+                override fun create(container: SdkContainer): UtilityAgent = UtilityAgent(
+                    container.getContextManager(),
+                    container.getDirectiveSequencer()
+                )
+            })
         }
         .build()