+++ nugu-client-transport-http2/src/main/java/com/skt/nugu/sdk/client/port/transport/http2/utils/BackOff.kt
@@ -137,7 +137,6 @@ class BackOff private constructor(builder: Builder) {
     private fun isRetryableServiceException(code: Status.Code): Boolean {
         return when(code) {
             Status.Code.OK,
-            Status.Code.PERMISSION_DENIED,
             Status.Code.UNAUTHENTICATED -> false
             else -> true
         }